调整方法

+++ common/src/main/java/com/leo/commonutil/storage/IOUtil.java
@@ -3,7 +3,6 @@ package com.leo.commonutil.storage;
 import java.io.ByteArrayOutputStream;
 import java.io.Closeable;
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.concurrent.ExecutorService;
@@ -19,7 +18,13 @@ public class IOUtil {
         return mThreadPool;
     }
 
-    //从流中读取数据
+    /**
+     * 从流中读取数据
+     *
+     * @param inStream
+     * @return
+     * @throws Exception
+     */
     public static byte[] read(InputStream inStream) throws Exception {
         ByteArrayOutputStream outStream = new ByteArrayOutputStream();
         byte[] buffer = new byte[1024];
@@ -35,14 +40,16 @@ public class IOUtil {
      * Close closable object and wrap {@link IOException} with {@link
      * RuntimeException}
      *
-     * @param closeable closeable object
+     * @param closeables closeable object
      */
-    public static void close(Closeable closeable) {
-        if (closeable != null) {
-            try {
-                closeable.close();
-            } catch (IOException e) {
-                throw new RuntimeException("IOException occurred. ", e);
+    public static void close(Closeable... closeables) {
+        for (Closeable closeable : closeables) {
+            if (closeable != null) {
+                try {
+                    closeable.close();
+                } catch (IOException e) {
+                    throw new RuntimeException("IOException occurred. ", e);
+                }
             }
         }
     }
@@ -51,60 +58,32 @@ public class IOUtil {
      * Close closable object and wrap {@link IOException} with {@link
      * RuntimeException}
      *
-     * @param cursor closeable object
+     * @param cursors closeable object
      */
-    public static void closeQuietly(Closeable cursor) {
-        try {
-            if (null != cursor) {
-                cursor.close();
+    public static void closeQuietly(Closeable... cursors) {
+        for (Closeable cursor : cursors) {
+            try {
+                if (null != cursor) {
+                    cursor.close();
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
             }
-        } catch (Exception e) {
-            e.printStackTrace();
         }
     }
-
-    /**
-     * 保存文本
-     *
-     * @param fileName 文件名字
-     * @param content  内容
-     * @param append   是否累加
-     * @return 是否成功
-     */
-    public static void saveTextValue(String fileName, String content, boolean append) {
-        getThreadPool().execute(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    File textFile = new File(fileName);
-                    if (!append && textFile.exists()) {
-                        textFile.delete();
-                    }
-
-                    FileOutputStream os = new FileOutputStream(textFile);
-                    os.write(content.getBytes("UTF-8"));
-                    os.close();
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-    }
-
 
     /**
      * 删除目录下所有文件
      *
      * @param Path 路径
      */
-    public static void deleteAllFile(String Path) {
-        // 删除目录下所有文件
+    public static void delAllFile(String Path) {
         File path = new File(Path);
         File files[] = path.listFiles();
-        if (files != null) {
+        if (files != null && files.length != 0) {
             for (File tfi : files) {
                 if (tfi.isDirectory()) {
-                    System.out.println(tfi.getName());
+                    delAllFile(tfi.getPath());
                 } else {
                     tfi.delete();
                 }
